@@ -1,4 +1,4 @@
-import { Inject, Injectable } from '@angular/core';
+import { Inject, Injectable, inject, Injector, DestroyRef, assertInInjectionContext } from '@angular/core';
 import { Router } from '@angular/router';
 import { BindQueryParamsManager } from './bind-query-params-manager';
 import { BIND_QUERY_PARAMS_OPTIONS } from './options';
@@ -6,9 +6,16 @@ import { BindQueryParamsOptions, CreateOptions, QueryDefOptions } from './types'
 
 @Injectable({ providedIn: 'root' })
 export class BindQueryParamsFactory {
-  constructor(private router: Router, @Inject(BIND_QUERY_PARAMS_OPTIONS) private options: BindQueryParamsOptions) {}
+  private router = inject(Router);
+  private options = inject(BIND_QUERY_PARAMS_OPTIONS);
 
-  create<T>(defs: QueryDefOptions<T>[] | QueryDefOptions<T>, createOptions?: CreateOptions): BindQueryParamsManager<T> {
-    return new BindQueryParamsManager<T>(this.router, defs, this.options, createOptions);
+  create<T>(defs: QueryDefOptions<T>[] | QueryDefOptions<T>, options?: CreateOptions): BindQueryParamsManager<T> {
+    !options?.injector && assertInInjectionContext(this.create);
+    const injector = options?.injector ?? inject(Injector);
+    const manager = new BindQueryParamsManager<T>(this.router, defs, this.options, options);
+
+    injector.get(DestroyRef).onDestroy(() => manager.destroy());
+
+    return manager;
   }
 }
